+++ changelog.md
@@ -1,3 +1,14 @@
+# 1.1.0
+### Additions
+- Added Silver Bulb (and all its oxidized/waxed variants)
+
+### Changes
+- Renamed all "Oxidized" silver blocks to "Tarnished"
+
+### Fixes
+- Fixed Bolt and Flow armor trims not working with silver as a material
+
+
 # 1.0.1
 ### Fixes
 - Fixed leather armor items not displaying correctly in the inventory.